moon_manager/tests/unit_python/requirements.txt
hug!=2.5.0
moon_utilities
sqlalchemy
pymysql
pytest-mock
pytest-cov
requests_mock
tinydb
pytest
pytest-benchmark
pyaml
